// src/app/shared/interfaces/icones.model.ts
export interface IconesMV {
  id?: any;
  svg?: string;
  label?: string;
  function?: any;
  info?: boolean;
  cssClass?: string;
  isExecutado?: boolean;
  isProximo?: boolean;
  isPendente?: boolean;
  proximo?: boolean;
  error?: boolean;
  infoLinha1?: any;
  infoLinha2?: any;
  iconColor?: any;
  lineColor?: any;
  style?: any;
}